Add bus schedule ui (#9)

+++ core/model/src/commonMain/kotlin/com/alad1nks/dubovozki/core/model/Bus.kt
@@ -3,7 +3,7 @@ package com.alad1nks.dubovozki.core.model
 data class Bus(
     val id: Int,
     val dayOfWeek: DayOfWeek,
-    val dayTime: Long,
+    val dayTime: Int,
     val dayTimeString: String,
     val station: Station,
     val direction: Direction,
